+++ blocks/article-hero/article-hero.css
@@ -8,12 +8,6 @@
     }
 }
 
-@media screen and (min-width: 600px) and  (max-width: 1260px){
-  .hero-collection {
-    height: 37vh;
-  }
-}
-
 .article-hero .hero-title {
     letter-spacing: 3px;
     text-transform: uppercase;
@@ -23,12 +17,12 @@
     width: 100%;
     height: 260px;
     object-fit: cover;
-    filter: brightness(70%);
 }
 
 @media screen and (min-width: 600px) {
     .article-hero img {
         height: 100%;
+        filter: brightness(70%);
     }
 }
 
@@ -69,42 +63,3 @@
         background-color: rgb(255 255 255 / 80%);
     }
 }
-
-  .article-hero .hero-collection-overlay {
-    position: absolute;
-    top: 75%;
-    left: 35%;
-    color: #fff;
-    font-size:40px;
-    margin-bottom: 8px;
-  }
-
-  .article-hero h1.hero-collection-title{
-    font-size: 40px;
-    text-transform: uppercase;
-    letter-spacing: 1px;
-  }
-
-  .article-hero p.hero-collection-subtitle{
-    font-size: 16px;
-    letter-spacing: 1px;
-    line-height: 28px;
-  }
-
-@media (max-width: 1260px){
-  .article-hero .hero-collection-overlay {
-    top: 20%;
-    left: 3%;
-  }
-
-  .article-hero h1.hero-collection-title{
-    font-size: 23px;
-  }
-
-  .article-hero p.hero-collection-subtitle{
-    font-size: 14px;
-    line-height: 14px;
-  }
-}
-
-
