+++ blocks/collection-hero/collection-hero.css
@@ -1,0 +1,82 @@
+.collection-hero {
+    position: relative;
+}
+
+@media screen and (min-width: 1200px) {
+    .collection-hero {
+        height: 100vh;
+        padding-bottom: 50px;
+    }
+}
+
+.collection-hero img {
+    width: 100%;
+    height: 260px;
+    object-fit: cover;
+    filter: brightness(70%);
+}
+
+@media screen and (min-width: 600px) {
+    .collection-hero img {
+        height: 100%;
+    }
+}
+
+.collection-hero div.hero-description {
+    max-width: 100%;
+    margin: auto;
+}
+
+.collection-hero div.hero-description > p {
+   text-align: center;
+}
+
+.section.collection-hero-container {
+    max-width: unset;
+    padding: unset;
+}
+
+.section > div.collection-hero-wrapper {
+    max-width: unset;
+}
+
+.collection-hero .hero-collection-overlay {
+  position: absolute;
+  top: 20%;
+  left: 5%;
+  color: #fff;
+  font-size:40px;
+  font-weight: 400;
+  margin-bottom: 8px;
+}
+
+  .collection-hero h1.hero-collection-title{
+    font-size: 26px;
+    font-weight: 400;
+    text-transform: uppercase;
+    letter-spacing: 1px;
+  }
+
+  .collection-hero p.hero-collection-subtitle{
+    font-size: 16px;
+    letter-spacing: 1px;
+    line-height: 28px;
+  }
+
+@media (min-width: 1200px){
+  .collection-hero .hero-collection-overlay {
+    top: 68%;
+    left: 15%;
+  }
+
+  .collection-hero h1.hero-collection-title{
+    font-size: 40px;
+  }
+
+  .collection-hero p.hero-collection-subtitle{
+    font-size: 14px;
+    line-height: 14px;
+  }
+}
+
+
